moved primitive generator
